+++ spec/models/concerns/events/event_index_handler_spec.rb
@@ -2,7 +2,7 @@
 
 require "rails_helper"
 
-RSpec.describe(EventIndexHandler, type: :concern) do
+RSpec.describe(Events::EventIndexHandler, type: :concern) do
   include ActiveSupport::Testing::TimeHelpers
 
   let(:event) { build(:event) }
